Conflicts:
	src/lib/Supra/Controller/Pages/Listener/EntityAuditListener.php

+++ src/webroot/cms/content-manager/pagecontent/PagecontentAction.php
@@ -14,6 +14,7 @@ use Supra\Editable;
 use Supra\Controller\Pages\Event\PageEventArgs;
 use Supra\Controller\Pages\Event\AuditEvents;
 use Supra\ObjectRepository\ObjectRepository;
+use Supra\Controller\Pages\Configuration\BlockControllerConfiguration;
 
 /**
  * Controller for page content requests
@@ -127,8 +128,10 @@ class PagecontentAction extends PageManagerAction
 			$property = $blockController->getProperty($propertyName);
 
 			// Could be new, should persist
-			$this->entityManager->persist($property);
-			/* @var $property Entity\BlockProperty */
+			if ( ! $property instanceof Entity\SharedBlockProperty) {
+				$this->entityManager->persist($property);
+				/* @var $property Entity\BlockProperty */
+			}
 
 			$editable = $property->getEditable();
 
@@ -253,7 +256,10 @@ class PagecontentAction extends PageManagerAction
 		$this->entityManager->flush();
 		
 		$controllerClass = $this->getPageControllerClass();
-		
+
+		// Regenerate the request object
+		$request = $this->getPageRequest();
+
 		// Need to be inside page and block controller scopes
 		ObjectRepository::beginControllerContext($controllerClass);
 		ObjectRepository::beginControllerContext($blockController);
@@ -293,13 +299,14 @@ class PagecontentAction extends PageManagerAction
 		$this->checkLock();
 		
 		$blockId = $this->getRequestParameter('block_id');
-		
-		$blockEntity = Entity\Abstraction\Block::CN();
-		$blockQuery = $this->entityManager->createQuery("SELECT b FROM $blockEntity b
-					WHERE b.id = ?0");
-		
-		$blockQuery->setParameters(array($blockId));
-		$block = $blockQuery->getSingleResult();
+
+		$block = $this->entityManager->find(Entity\Abstraction\Block::CN(), $blockId);
+
+		if (empty($block)) {
+			throw new CmsException(null, 'Block was not found');
+		}
+
+		$this->checkBlockSharedProperties($block);
 		
 		$this->entityManager->remove($block);
 		$this->entityManager->flush();
@@ -308,6 +315,35 @@ class PagecontentAction extends PageManagerAction
 		
 		// OK response
 		$this->getResponse()->setResponseData(true);
+	}
+
+	/**
+	 * Will confirm the removal if shared properties exist
+	 * @param Entity\Abstraction\Block $block
+	 */
+	private function checkBlockSharedProperties(Entity\Abstraction\Block $block)
+	{
+		$class = $block->getComponentClass();
+		$configuration = ObjectRepository::getComponentConfiguration($class);
+
+		$hasSharedProperties = false;
+
+		// Collects all shared properties
+		if ($configuration instanceof BlockControllerConfiguration) {
+			foreach ($configuration->properties as $property) {
+				/* @var $property BlockPropertyConfiguration */
+				if ($property->shared) {
+					$hasSharedProperties = true;
+
+					// enough to find one
+					break;
+				}
+			}
+		}
+
+		if ($hasSharedProperties) {
+			$this->getConfirmation("{#page.delete_block_shared_confirmation#}");
+		}
 	}
 	
 	/**
